Provide `LLVMTypeWrapper` trait
Instead of keeping raw pointer fields (of types like `LLVMValueRef`,
`LLVMMetadataRef`) public and defining constructors with different names,
provide the `LLVMTypeWrapper` trait with `from_ptr()` `as_ptr()` method
for all of them.

This will allow to convert all safe wrappers from and to raw pointers
with one method, which is going to be helpful for building macros for
them.