🔒🌯 Provide `LLVMTypeWrapper` trait, add more wrappers
Instead of keeping raw pointer fields (of types like `LLVMValueRef`,
`LLVMMetadataRef`) public and defining constructors with different
names, provide the `LLVMTypeWrapper` trait with `from_ptr()` and
`as_ptr()` methods.

This will allow to convert all safe wrappers from and to raw pointers
with one method, which is going to be helpful for building macros for
them.

On top of that, add more wrappers:

* `Module`
* `BasicBlock`
* `GlobalAlias`
* `GlobalVariable`
* `Argument`

Use these wrappers in iterators, make sure they don't expose the raw
pointers to the callers.